+++ src/main/resources/application-jpa.properties
@@ -17,7 +17,7 @@ spring.jpa.properties.hibernate.format_sql=true
 logging.level.root=WARN
 logging.level.goorm.mybatisboard=DEBUG
 logging.level.goorm.mybatisboard.post.controller.PostController=INFO
-logging.level.goorm.mybatisboard.post.service.PostService=DEBUG
+logging.level.goorm.mybatisboard.post.service.MyBatisPostServiceImpl=DEBUG
 logging.level.goorm.mybatisboard.post.mapper.PostMapper=DEBUG
 logging.level.goorm.mybatisboard.post=DEBUG
 logging.level.org.springframework.web=INFO
