+++ src/main/resources/application.properties
@@ -33,7 +33,7 @@ mybatis.configuration.map-underscore-to-camel-case=true
 logging.level.root=INFO
 logging.level.goorm.mybatisboard=DEBUG
 logging.level.goorm.mybatisboard.post.controller.PostController=INFO
-logging.level.goorm.mybatisboard.post.service.PostService=DEBUG
+logging.level.goorm.mybatisboard.post.service.MyBatisPostServiceImpl=DEBUG
 logging.level.goorm.mybatisboard.post.mapper.PostMapper=DEBUG
 logging.level.goorm.mybatisboard.post=DEBUG
 logging.level.org.springframework.web=INFO
